Restructured directories to represent MVC pattern
