+++ src/styles/style.sass
@@ -1,14 +1,14 @@
 @font-face
   font-family: fredokaOne
-  src: url('./FredokaOne-Regular.ttf')
+  src: url('../assets/FredokaOne-Regular.ttf')
 
 @font-face 
   font-family: ebGaramond
-  src: url('./EBGaramond-VariableFont_wght.ttf')
+  src: url('../assets/EBGaramond-VariableFont_wght.ttf')
 
 @font-face
   font-family: openSans
-  src: url('./OpenSans-VariableFont_wdth,wght.ttf')
+  src: url('../assets/OpenSans-VariableFont_wdth,wght.ttf')
 
 //#7db6d6, #0071af, #5ca4cc, #00659d, #006095, #59cea7, #31c292
 //#ec9090, #dfdfdf, #bebebe, 
@@ -271,9 +271,6 @@ label[for="create-project"]
 // project-3dot-icon:after
 //   top: 80px
 
-
-
-
 .project-more-options 
   width: 46px
   text-align: center
@@ -426,7 +423,6 @@ textarea
   border-radius: 10px
   font-size: 20px
   color: #000033
-
 
 #create-task-button, #update-task-button
   margin-bottom: 0
@@ -624,4 +620,3 @@ footer > span
 
 .github-logo:hover
   transform: rotate(360deg) scale(1.2)
-
